requirements.txt
django
textfsm
djangorestframework